Fixed bug with MoveCursorToIndex.
An extra space was being added in some cases, which should have been
handled specifically by the mention results code instead of
MoveCursorToIndex.

A small UI bug still exists but seems to be an underlying issue with
tview. To Replicate:

Mention a name with no other text to the right.
Moving the cursor one index to the left will be an unhighlighted space.