Extract and use ChatViewInterface from ui.ChatView
To ease the path to implementing a new chat view, we extract the
interface with the existing chat view has established. We then have its
collaborators rely on the interview instead of the concret type.

The new chat view will implement this interface, so that collaborators
don't need to know anything has changed.