Fixed potential deadlock during channel loading
Due to the fact that channels where loaded in a background thread which
then moved 99% of it's logic back onto the UI-thread and waited for it
to finish, switching channels very quickly (Alt+L - Load previous
channel) caused a deadlock. The reasoning behind this back then, was to
reduce timeframes in which the UI doesn't respond, but it was basically
incorrect either way.