Improve editor event handling
The editor now always handles the externally set input handlers first.
This was changed due to incorrect values being returned from the input
handler, which in turn caused the upcoming draw to be skipped. While
this didn't cause any critical problems, it made the UI feel laggy at
times. Technically certain editor shortcuts could now be overwritten
from outside, but if the user choses to do so, we shall not care.