Wire event handling through the Window type
Previously, the `tview.Application` instance would bubble events down to
the focussed primitive first. This allowed the composed components to
handle events as they saw fit, but it didn't give the window a chance to
handle them.

To address this, the `windowman.Window` interface now has a
`HandleKeyEvent` event method, which is called by
`windowman.WindowManager` to begin bubbling the events down.

`windowman.Window` implementations are expected to track the currently
focussed component, so that they can bubble events down to it.

In the future, some of the event plumbing may be extractable to a base
type.